dc cleanup

+++ pychron/lasers/points/maker.py
@@ -475,7 +475,7 @@ class GridOverlay(AbstractOverlay):
         with gc:
             comp = self.component
             gc.clip_to_rect(
-                *(int(p) for p in (comp.x, comp.y, comp.width, comp.height))
+                *(float(p) for p in (comp.x, comp.y, comp.width, comp.height))
             )
             pos = comp.get_offset_stage_screen_position()
